Changes suggested project layout folder from **scripts** to **code**
Referenced in issue /issues/471, the *Before we start* episode 0 suggests keeping Python code in a **scripts** folder.
Unfortunately, Anaconda has a built-in `scripts` module that will shadow the student's **scripts** folder and make
```
import scripts.example
```
fail.

The issue suggests adding an __init__.py to **scripts**.
I believe that adds unnecessary complexity to the introductory material
and propose renaming the folder with this PR.